implemented integration test

+++ src/main/java/com/user/api/repository/UserRepository.java
@@ -2,11 +2,10 @@ package com.user.api.repository;
 
 import com.user.api.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByEmail(String email);
